requirements.txt
# Core dependencies with pinned versions for reproducibility
numpy==1.26.4
pandas==2.2.0
matplotlib==3.8.3
PyYAML==6.0.1
torch==2.2.1
